Moves devtools into its own cog.
